backend/requirements.txt
asgiref==3.7.2
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.3
pillow
djoser
reportlab